change type to decimal instead of int

+++ ascender/Dto/EntryDto.cs
@@ -3,5 +3,5 @@ namespace ascender.Dto;
 public class EntryDto
 {
     public string Name { get; init; }
-    public int Value { get; init; }
+    public decimal Value { get; init; }
 }
